refactor: type the snapshot model and derive TrendSnapshotView
